+++ src/com/nativebuild/util/misc/NativeBuildException.kt
@@ -1,0 +1,3 @@
+package com.nativebuild.util.misc
+
+open class NativeBuildException(message: String): Exception(message)